pyproject.toml
[project]
name = "llamaindex-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform>=1.90.0",
    "llama-index>=0.12.33",
    "llama-index-embeddings-google-genai>=0.1.0",
    "llama-index-llms-google-genai>=0.1.7",
    "llama-index-llms-vertex>=0.4.6",
    "llama-index-vector-stores-docarray>=0.3.0",
    "python-dotenv>=1.1.0",
]
